backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.1.0
Pillow==10.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
